// packages/contract-helpers/src/synthetix-contract/typechain/ISynthetix.d.ts
/* Autogenerated file. Do not edit manually. */
/* eslint-disable */

import {
  ethers,
  EventFilter,
  Signer,
  BigNumber,
  BigNumberish,
  PopulatedTransaction,
} from 'ethers';
import {
  Contract,
  ContractTransaction,
  CallOverrides,
} from '@ethersproject/contracts';
import { BytesLike } from '@ethersproject/bytes';
import { Listener, Provider } from '@ethersproject/providers';
import { FunctionFragment, EventFragment, Result } from '@ethersproject/abi';

interface ISynthetixInterface extends ethers.utils.Interface {
  functions: {
    'transferableSynthetix(address)': FunctionFragment;
  };

  encodeFunctionData(
    functionFragment: 'transferableSynthetix',
    values: [string],
  ): string;

  decodeFunctionResult(
    functionFragment: 'transferableSynthetix',
    data: BytesLike,
  ): Result;

  events: {};
}

export class ISynthetix extends Contract {
  connect(signerOrProvider: Signer | Provider | string): this;
  attach(addressOrName: string): this;
  deployed(): Promise<this>;

  on(event: EventFilter | string, listener: Listener): this;
  once(event: EventFilter | string, listener: Listener): this;
  addListener(eventName: EventFilter | string, listener: Listener): this;
  removeAllListeners(eventName: EventFilter | string): this;
  removeListener(eventName: any, listener: Listener): this;

  interface: ISynthetixInterface;

  functions: {
    transferableSynthetix(
      account: string,
      overrides?: CallOverrides,
    ): Promise<{
      transferable: BigNumber;
      0: BigNumber;
    }>;

    'transferableSynthetix(address)'(
      account: string,
      overrides?: CallOverrides,
    ): Promise<{
      transferable: BigNumber;
      0: BigNumber;
    }>;
  };

  transferableSynthetix(
    account: string,
    overrides?: CallOverrides,
  ): Promise<BigNumber>;

  'transferableSynthetix(address)'(
    account: string,
    overrides?: CallOverrides,
  ): Promise<BigNumber>;

  callStatic: {
    transferableSynthetix(
      account: string,
      overrides?: CallOverrides,
    ): Promise<BigNumber>;

    'transferableSynthetix(address)'(
      account: string,
      overrides?: CallOverrides,
    ): Promise<BigNumber>;
  };

  filters: {};

  estimateGas: {
    transferableSynthetix(
      account: string,
      overrides?: CallOverrides,
    ): Promise<BigNumber>;

    'transferableSynthetix(address)'(
      account: string,
      overrides?: CallOverrides,
    ): Promise<BigNumber>;
  };

  populateTransaction: {
    transferableSynthetix(
      account: string,
      overrides?: CallOverrides,
    ): Promise<PopulatedTransaction>;

    'transferableSynthetix(address)'(
      account: string,
      overrides?: CallOverrides,
    ): Promise<PopulatedTransaction>;
  };
}